PaddedBatch.seq_len_mask type fix

+++ dltranz/trx_encoder.py
@@ -41,7 +41,10 @@ class PaddedBatch:
             B, T = next(iter(self._payload.values())).size()
         else:
             B, T, _ = self._payload.size()
-        return (1 - torch.triu(torch.ones(T, T, device=self._length.device, dtype=torch.int64), 1))[self._length - 1]
+        return (1 - torch.triu(
+            torch.ones(T, T, device=self._length.device, dtype=torch.int64),
+            1
+        ))[self._length.long() - 1]
 
 
 class NoisyEmbedding(nn.Embedding):
